openvswitch robot: update the job with fixes
The job will now apply the patch first, and then execute checkpatch. This
allows checkpatch fixes to not flag themselves.

Additionally, switch to using the jq utility where possible.  It is a
better json parser than egrep.

Track message subjects in addition to the message IDs, and reports that
back to the list.

Finally, hooked up with make dist-check so that the dist check suite
will be executed.

Signed-off-by: Aaron Conole <aconole@redhat.com>