pw-ci: include instance information
The series IDs are not good enough to track, since they could be
reused across patchwork instances.  It isn't even good enough to
try and select on project+id - one could envision a chance that
they collide.

Introduce the ability to filter on series_instance, which is
unique for all projects on an instance.

Most of the information was already available, so might as well
just pull it.

This DOES make a slight re-query, since we will re-try with older
dates than usual (unless the query is weekly), but that will not do
anything if the series ID already exists.

Signed-off-by: Aaron Conole <aconole@redhat.com>