pw-ci: allow for downloads to fail
Periodically, a patchwork instance might receive enough traffic that a
URL get fails.  This happens occasionally with a running instance of
the 0-day bot.  When it occurs, we need to re-submit the series.

This allows that to happen in a way that will continuously re-try without
blocking forward progress by constantly allowing a complete batch to
proceed.

Signed-off-by: Aaron Conole <aconole@redhat.com>