pw_mon: clear poll flag for old series in the DB
Series will possibly pile up if the builds associated were
interrupted or otherwise altered.  This has the disadvantage
of causing the system to continually poll for build status
even on builds that were intentionally cancelled.  If a
patch is superseded, rejected, or accepted, we don't
even need to care any longer.

Signed-off-by: Aaron Conole <aconole@redhat.com>