pw_mon: convert to unversioned API
The best way to access the newest APIs from patchwork is to
use the unversioned API.  Older versions may eventually become
deprecated, and we would have to decide on the 'best' version
to use.  The unversioned API will always be available for
us to use, and will prefer the newest versions.

Acked-by: Michael Santana <msantana@redhat.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>