ci_mon: allow skipping by project id
Most of the infrastructure uses both patchwork project and
patchwork instance to discriminate on series information.
The monitor scripts currently run for all projects.  This
means we might send email to the wrong lists if a ci_mon
report needs to go to different lists for different projects.

Correct this by adding a parameter for patchwork project
data that will be used if present to skip a series.

Acked-by: Michael Santana <msantana@redhat.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>