github: support repositories with multiple workflows
The existing github monitor code only supports a single workflow job.
This worked as an initial implementation, but some projects are
configuring multiple workflows and the monitor is unable to cope with
this.

To resolve this, we pull the workflow configurations and then ensure
that all workflows are completed, and emit results once all have
finished.  Each workflow run can now emit two 'pw|...' strings.

As part of this change, we add a new field 'test_name to the 'pw|...'
string to support passing the workflow name around.  CI systems like
Travis won't report anything here so test_name will be blank for those
systems.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>
Acked-by: Michael Santana <msantana@redhat.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>