dpdk/config.xml: Resolve series dependency.
Proposed patches in DPDK may have dependency on other work in progress
series.
This is described in DPDK contributing guide [1].

For now, only handle a series dependency.

Trying to resolve series dependencies might be too complex, but we can
rely on the dependency previous submission being already referenced in
the upload repository.

1: https://doc.dpdk.org/guides/contributing/patches.html#patch-dependencies

Note: this is an untested patch so it is probably broken.
I am unsure how I could try it safely, but posting as a kind of RFC.

Signed-off-by: David Marchand <david.marchand@redhat.com>