pw_mon: switch to using the events API for tracking series
The Patchwork API implemented events long ago, and emits events on
major changes to patch series.  For example, when a series is
officially marked completed, a special event gets generated noting the
series marker.

When pw-ci was first implemented, such API and concepts weren't
available, so we had to work around it based on what existed.  This has
some deficits, the biggest is that the way patch dates are stored is
based on the patch noted date (rather than received date), and that can
cause us to miss patches in patchwork.

The Events API doesn't suffer from this issue.  It also means we no
longer will generate queries for whether a series is completed, since
we can just monitor the events API for those series-completed events.

The drawback here is that we will generate extra requests to the
server in the normal case.  That's because the events API doesn't
include some needed details about the series (like authorship details)
that we track.  But the upside is we shouldn't miss patches with
authorship dates in the past any longer.

Signed-off-by: Aaron Conole <aconole@redhat.com>
Acked-by: Michael Santana <msantana@redhat.com>