recheck: Add a recheck parser for patchwork comments
Add a recheck parsing tool that will allow for labs to build a
recheck workflow based on specific recheck labels and projects,
with an associated state machine and querying interface.

The output of the recheck parsing tool is json and can be fed to
jq or other json parsing utilities for better field support.

Acked-by: Michael Santana <msantana@redhat.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>