pw_mon: recheck via patchwork
Previous selection criteria for a patch series was based on
scanning open series in the series DB.  However, the series DB
only tracks those patches where a test is still outstanding.
Once a test is completed, it will be marked as inactive and
the scan won't find it.  This wasn't caught during testing
because the test setup had the series branch updates disabled
which caused a series to stay active even after the results
would be marked completed.

Check the actual patch series directly.  This will only work
for patchwork instances that don't have many (say only a few
dozen) outstanding patchwork series.  For heavily loaded
patchwork instances that keep hundreds or thousands of
series outstanding the only solution is to use the events
interface to poll.  That will be a forthcoming option.

The checking only currently supports selecting the first page
of results.  Support for pagination will be forthcoming.

Signed-off-by: Aaron Conole <aconole@redhat.com>