cirrus: Add an initial polling mechanism
Cirrus-CI is used to provide a testing environment for FreeBSD, among
others.  As of now, the ci monitoring framework doesn't support checking
for cirrus-ci status, but that can be changed.  We add an initial polling
similar to the original GitHub Actions or Travis-CI polling mechanisms
that merely provides pass/fail/warn with a URL.  This can be improved
later to provide logs on a per-task (similar to GHA Jobs) basis.

Signed-off-by: Aaron Conole <aconole@redhat.com>